Fixed create_app_test_http.ts and ran deno fmt on it
create_app_test_http.ts would get a nonexistant branch
on the repository 'drashland/deno-drash' in the case of
a forked repository. That is, the branch might exist on
the forked repository, but it would NOT exist on
'drashland/deno-drash'. This would result in an error in
the Github CI tests.