Added deprecation logger and @deprecation JS doc labels
I used the method addLogger to add a ConsoleLogger 'DeprecationLogger'
to the Loggers namespace in the constructor of TemplateEngine. To me,
this was the most logical place to call addLogger. I don't see
why we couldn't just add a deprecation log message at the beginning of
the constructor for TemplateEngine.