Trying to fix a test in mod_test.ts
Since I used the function addLogger, there was an extra logger
added to the Logger array (Drash.Loggers). One of the addLogger
tests in mod_test.ts added a testLogger, and compared the JS
object { 'TestLogger': testLogger } to Drash.Loggers. However,
a DeprecateLogger was added, making it NOT equal to
{ 'TestLogger': testLogger }.